+++ executor/insert_common.go
@@ -1074,9 +1074,8 @@ func (e *InsertRuntimeStat) String() string {
 	if e.Prefetch > 0 {
 		buf.WriteString(fmt.Sprintf("check_insert:{total_time:%v, mem_insert_time:%v, prefetch:%v", e.CheckInsertTime, e.CheckInsertTime-e.Prefetch, e.Prefetch))
 		if e.SnapshotRuntimeStats != nil {
-			rpcStr := e.SnapshotRuntimeStats.String()
-			if len(rpcStr) > 0 {
-				buf.WriteString(fmt.Sprintf(", rpc:{%s}", rpcStr))
+			if rpc := e.SnapshotRuntimeStats.String(); len(rpc) > 0 {
+				buf.WriteString(fmt.Sprintf(", rpc:{%s}", rpc))
 			}
 		}
 		buf.WriteString("}")
